Only try file removal if it exists

+++ app/admin/code.rb
@@ -31,7 +31,7 @@ ActiveAdmin.register Code do
     zipfile_path = "#{Rails.root}/tmp/#{zipfile_name}"
     image_names = []
 
-    File.delete(zipfile_path)
+    File.delete(zipfile_path) if File.exist?(zipfile_path)
 
     Code.all.each do |code|
       canvas = Magick::ImageList.new
